Move lang switch from Header.jsx to component LangSwitch.jsx
