+++ frontend/src/components/Header/Header.scss
@@ -55,21 +55,20 @@
     margin: 0;
     padding: 0;
     list-style-type: none;
+    align-items: center;
   }
 
   &__location-item {
-    margin-right: 15px;
-
+    margin-right: 30px;
+    &:last-child {
+      margin-right: 0;
+    }
     &_lang {
-      padding: 5px 15px;
+      padding: 2px;
       display: flex;
       align-items: center;
       background-color: #f3f3f3;
       border-radius: 25px;
-    }
-
-    &:last-child {
-      margin-right: 0;
     }
   }
 
@@ -101,8 +100,6 @@
 
   &__btn-sign {
     @include mainText;
-
-    padding: 5px 15px;
     display: flex;
     align-items: center;
     color: $black;
